+++ BackEnd/src/modules/notifications/notifications.module.ts
@@ -4,9 +4,17 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { NotificationsController } from './notifications.controller';
 import { NotificationsService } from './notifications.service';
 import { Notification } from './entities/notification.entity';
+import { NotificationPreference } from './entities/notification-preference.entity';
+import { NotificationLog } from './entities/notification-log.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Notification])],
+  imports: [
+    TypeOrmModule.forFeature([
+      Notification,
+      NotificationPreference,
+      NotificationLog,
+    ]),
+  ],
   providers: [NotificationsService],
   controllers: [NotificationsController],
   exports: [NotificationsService],